Move the DerivedFieldScript under the right package
